//实践证明int负数整数都表示非0——真

+++ Choice_Question/test.c
@@ -13,14 +13,14 @@
 
 int main()
 {
-    //测试double的能否成为判断while循环的数
-    double a = 1.11, b = -1.11;
+    //测试int的能否成为判断while循环的数
+    int a = 1.11, b = -1.11;
     int flag=4;
     while (flag != 0 && a){
         flag--;
         printf("1. Yes!\n");
     }
     for (flag = 4; flag >= 0 && b ; flag--) printf("2. Yes!\n");
-    //实践证明double负数整数都表示非0——真
+    //实践证明int负数整数都表示非0——真
     return 0;
 }
